src/order-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.8.2
httpx==0.27.0
pyjwt==2.8.0
aiokafka==0.10.0
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
redis==5.0.7
pybreaker==1.2.0
python-dotenv==1.0.1
loguru==0.7.2

